refactor: rename types.js file
